requirements.txt
# Basic Dependencies
numpy
jinja2
networkx

# Flask - Web Framework
Flask
Flask-Cors

# Environment Variables
python-dotenv

# PyTorch - Deep Learning Framework
torch

# WhisperX - Speech Recognition Model
whisperx

# Audio Processing
ffmpeg-python
pyannote.audio

# LangChain - LLM Framework
langchain
langchain-core
langchain-openai

# OpenAI - LLM Provider
openai

# Hugging Face - Required for WhisperX
huggingface-hub

# Dateparser
dateparser

#.\venv\Scripts\activate
# pip install -r requirements.txt